Framework support for multiple handler chains
Some commands (such as DATA) need to re-enter the handler processing
chain once they are completed. For DATA, the first SMTP reply is
issued immediately. After the message body is collected, a second
SMTP reply is issued.

To allow the implementation of separate handlers for each stage of the
DATA command, each stage is implemented as a separate chain. The early
stage is triggered by the DATA command, and the final stage is
triggered by the SCHS_CHAIN exit code of the first stage. The final
stage is run as a handler chain for the BODY command.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>