Change AUTH command handling architecture
It's now handled internally in the smtp server engine.

Transitioning to different stages of the AUTH is done by
using "pseudo" commands:
	* ALOU - AUTH LOGIN Username
	* ALOP - AUTH LOGIN Password
	* APLP - AUTH PLAIN Password

Currently only LOGIN and PLAIN AUTH schemes are supported.
Others, such as DIGEST-MD5, CRAM-MD5, etc are not yet
supported, but can be integrated on top of the new design.

The engine does base64 encoding/decoding using openssl lib
functions and it stores plain text auhentication information
(such as username, password and auth method) in the server's
context, allowing the sql logging module and other modules
to use that information.

The actual authentication to the real smtp server is done
by the mod_proxy ALOP and APLP handlers.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>