Tolerate extra params to "MAIL FROM"
Some versions of sendmail (8.13.8, for instance) add SIZE=nnn to the
"MAIL FROM" command. The information is probably useless, because we
cannot trust the client. We can at most reject the mail immediately if
the declared size is beyond a treshold, or stop accepting message body
data when we reach the declared size.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>