Added function string_kv_split()
This function takes a string in the format
key1=value1<separator> key2=value2<separator>, ...
and returns a list of struct kv_pair containing
the key and values. The function alters the
original string and makes sure to strip surrouning
whitespaces. It doesn't alter the whitespace inside
the keys or values.

Signed-off-by: Ionut Nicu <ionut.nicu@cloudbit.ro>