Reorganize SMTP server context implementation
At the JavaScript level, it's better to use a SmtpServer class (instead
of the single global object called smtpServer). When a new connection is
made to the server, a new SmtpServer instance is created (automatically,
by the C code). This instance holds all session-specific data (such as
the envelope sender, the recipients list, etc.) that was previously held
in smtpServer.session.

The SmtpClient instance (in the default proxy use case) will also be
held as a property of the SmtpServer instance.

This approach avoids holding any session-specific data in the global
context. In the future, it will allow to use multithreading (instead of
separate worker processes), where each thread holds its own SmtpServer
instance.

As a rule of thumb:
- All configuration data is stored as "static properties" in the
  SmtpServer "class". Actually, SmtpServer is a function object and
  configuration data is stored as regular properties in this object.
  This corresponds to the concept of "static properties" in the
  JavaScript object model.
- All session data is stored in SmtpServer instances.

Custom SMTP command handlers are defined in SmtpServer.prototype. This
way, they will be inherited by all SmtpServer instances.

At the C level, struct smtp_server_context should only contain C
specific data (which is not exposed in any way to JavaScript), such as
the buffered file descriptor of the client connection, a pointer to the
SmtpServer instance object, etc.

As the JavaScript implementation of session-specific data is reviewed
and cleaned up, the old C implementation (represented by specific fields
in struct smtp_server_context and the corresponding functions) will be
removed.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>