Rename SmtpResponse "messages" property to "message"
Using the plural is confusing because in fact the SMTP response contains
just one message. When the "message" property is an array, the message
is a multi-line message where each array element represents one line.
The lines are glued together with a '\n' character before they are sent
back to the client - see smtp_server_response().

These changes do NOT include inactive (commented out) code that needs to
be ported to Duktape, since that code will have to be rewritten anyway.