conf: Remove obsolete JS handler examples
This code was added before the first JS integration (using mozjs) to
serve as an example of what the SMTP server API for JS would look like.
It has been commented out since the beginning, and now that the JS
integration is complete (and even ported to a new JS engine), the
example code is obsolete.

This commit removes the obsolete example code. The parts that store the
SMTP transaction details into a SQL database are still preserved but
moved to the actual JS handlers that use the current API. The database
related code is still commented out because the libjssql integration has
not been done yet.