Change bfd functions return value and drop bfd_seek()
Change the bfd functions to return 0 on success and a (positive) POSIX
error code on error (instead of -1 and setting errno). For functions
that need to return a value, such as the number of bytes read/written,
return the POSIX error code as a negative number instead.

This change makes the code more consistent with the SMTP code, where
error codes are returned through the function return value. It also
prepares the bfd code for the upcoming TLS support.

While at it, drop the bfd_seek() function, which is only used by the
pexec code. The pexec code is currently disabled and will have to be
revisited anyway. The temporary file that stores the message body is
now closed by the SMTP server code before handlers are called, so the
pexec code will have to open a new file descriptor anyway, and the
cursor will be automatically positioned at the beginning of the file.