Add STARTTLS support using OpenSSL
The global SSL context is initialized in src/mailfilter.c. The private
key and the certificate chain are loaded from files. The file paths are
passed as command line parameters. If one of the parameters is missing
or if there is some error reading the files and initializing the SSL
context, Mailfilter still starts but STARTTLS is disabled.

The global SSL context initialization is done very early, immediately
after the command line parameters are inspected. Notably, it is done
before other potentially dangerous (from a security perspective)
actions, such as starting the JS engine and running the JS config.
The plan is to add code in a future commit to drop privileges after
initializing the SSL context.

The bulk of the STARTTLS logic is implemented in src/smtp_server.c.
The data I/O interface to OpenSSL is conveniently implemented in the
buffered file descriptor layer (src/bfd.c). This makes it possible to
make it all transparent to the existing data I/O code and keep the
changes to a minimum.

Resolves: #20