Differentiate the SMTP client and server hostname
Currently Mailfilter assumes the client hostname (specified through
HELO or EHLO) as its own hostname. This is reflected in the hostname that
is sent to the backend server in HELO/EHLO commands, and the `Received`
header that is added to the email before it's forwarded to the backend
server. This is confusing and likely not standard compliant either, and
there is no real benefit to it.

Move the `hostname` property to the SmtpServer prototype and initialize
it to the system hostname. The JS environment can overwrite it in the
global scope easily if that's needed. Use a new `clientname` property
that reflects the HELO/EHLO hostname received from the client. This is
an instance property, initialized to null by the SmtpServer constructor.