Add a message size limit to the SpamAssassin client
The default is 2 MiB. Larger messages are not scanned (no connection is
made to spamd) and SpamAssassin::scan() returns null. The message size
is not limited by spamd itself but spamc implements a similar mechanism
to limit the message size. The concern is that spamd would use a huge
amount of memory for large messages. And spam is typically small.

The limit is pre-defined as SpamAssassin.MAX_MESSAGE_LEN and can be
changed by the JS configuration either globally at the prototype level
or for every SpamAssassin object instance.