Rename mod_spamassassin to mod_extscan
This prepares the ground for reimplementing the ClamAV scan interface.
It will be implemented using a direct TCP connection to clamd, and if
mod_clamav was a separate module, a lot of code would be duplicated.

There are no dependencies on external libraries, so it makes sense to
bring both scanners (SpamAssassin and ClamAV) into the same Mailfilter
module. That way, common code can be refactored into functions that can
be used in common by both scanners.