Remove the pexec compilation unit
Now that the SpamAssassin and ClamAV clients have been reimplemented
using direct TCP connections to the services, the pexec compilation unit
is obsolete. Besides, it uses the old mozjs API, and the code that
closes all open file descriptors before fork() (based on RLIMIT_NOFILE)
is highly inefficient on Linux.

For those reasons, remove the pexec compilation unit altogether. If
something similar is needed again in the future, it will have to be
reimplemented from scratch anyway.