Use a different tag if a virus is found
Emails that contain a virus pose an immediate thread and must be
classified differently from ordinary spam. At the very least, end
users must be aware of the threat by looking at the subject line.

This applies to the filter.js example as well if filters are
configured to be bypassed for the recipient (tagging only).

To be able to classify emails correctly, always scan for viruses
first because spam tests typically short-circuit further tests.