Do not truncate log messages for SMTP responses
When debug logging is enabled, Mailfilter logs the full SMTP dialogue
(commands and responses). In the initial implementation, the text part
of the response was truncated to a hard-coded limit of 50 characters
before being logged. As it turns out, the limit is too small, and
something like this gets truncated:

    4.7.1 <test.account@superlongdomain.com>: Relay access denied

There is no good reason to truncate these messages, so remove the
limit and truncation altogether.