Remove Sys.loadModule("mod_smtp_client.so") from config
Currently Sys.loadModule() is just a stub and doesn't do anything other
than just log a message. The SMTP client code does not depend on
external libraries and is compiled into the Mailfilter core. It will not
be implemented as a separate module even when module loading support is
eventually added.